chore(core): add query pause mechanism to be used for cold storage (#2781)

+++ core/src/main/java/io/questdb/cairo/O3Utils.java
@@ -76,7 +76,7 @@ public class O3Utils {
         }
     }
 
-    static void close(FilesFacade ff, long fd) {
+    static void close(FilesFacade ff, int fd) {
         if (fd > 0) {
             LOG.debug().$("closed [fd=").$(fd).$(']').$();
             ff.close(fd);
@@ -116,7 +116,7 @@ public class O3Utils {
         }
     }
 
-    static void unmapAndClose(FilesFacade ff, long dstFixFd, long dstFixAddr, long dstFixSize) {
+    static void unmapAndClose(FilesFacade ff, int dstFixFd, long dstFixAddr, long dstFixSize) {
         unmap(ff, dstFixAddr, dstFixSize);
         close(ff, dstFixFd);
     }
