+++ core/src/main/java/io/questdb/network/MutableIOContextFactory.java
@@ -33,7 +33,7 @@ import io.questdb.std.WeakMutableObjectPool;
 import java.io.Closeable;
 
 public class MutableIOContextFactory<C extends MutableIOContext<C>>
-        implements IOContextFactory<C>, Closeable, EagerThreadSetup {
+    implements IOContextFactory<C>, Closeable, EagerThreadSetup {
 
     private final ThreadLocal<WeakMutableObjectPool<C>> contextPool;
     private volatile boolean closed = false;
@@ -64,7 +64,7 @@ public class MutableIOContextFactory<C extends MutableIOContext<C>>
     }
 
     @Override
-    public C newInstance(long fd, IODispatcher<C> dispatcher) {
+    public C newInstance(int fd, IODispatcher<C> dispatcher) {
         return contextPool.get().pop().of(fd, dispatcher);
     }
